Multilevel Queue
Here  the  processes are permanently assigned to queue on entry and the processes do not move between queues. This method has the advantage of low scheduling overhead but is also not flexible which is sited as a disadvantage. 
However Multilevel Feedback Queue Scheduling allows a process to move between Queues so that a process does use too much of CPU time , this method is achieved by the idea to separate processes with different CPU-burst characteristics. . Similarly a process that waits too long in a lower-priority queue may be moved to a higher-priority queue. This form of aging prevents Starvation.
So we can say that Multilevel Feedback Scheduling is a method to prevent starving by allowing the processes to move freely between Queues.